Cache use cleanup.
Shiro since release 1.2.0 supports
authc and authz caching, that we use
instead of the removed (removed when
upped Shiro version to 1.2.0) homegrown
caching.

But, authc caching is disabled by default
in Shiro, and needs to be explicitly enabled.

This is done in XmlAuthenticatingRealm,
AbstractLdapAuthenticatingRealm and KenaiRealm,
as those were NOT performing authc caching
at all.

authz caching is enabled by default in Shiro,
still added explicit settings to make it clear.

Finally, XmlAuthenticatingRealm made
AuthenticatingRealm and originally it was
AuthorizingRealm without any purpose (as the
authz is not performed by it and method was
simply overridden).

This also means, that there was no authc caching
happening nor in XML realm, nor in any of the
LDAP realms (OSS or Pro).

Also, this fixes the misleading cache creation
of unused caches by some realms (ie. XmlAuthenticatingRealm
was creating authorizationCache but it never used it
as it never performed authz).